backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for the Winspotter project."
readme = "README.md"
requires-python = "~=3.13.0"
dependencies = [
    "boto3~=1.40.49",
    "fastapi~=0.118.0",
    "mangum~=0.19.0",
    "pwdlib[argon2]~=0.2.1",
    "pydantic-settings~=2.11.0",
    "pydantic[email]~=2.12.0",
    "pyjwt~=2.10.1",
    "python-multipart~=0.0.20",
]

[dependency-groups]
dev = [
    "ruff~=0.14.0",
    "uvicorn~=0.37.0",
]
